backend/requirements.txt
fastapi
uvicorn
python-docx
openai
python-multipart
